SDL3 port (#254)
* Updates for SDL3 support

Port of Pragtical to SDL3 using a cherry-picked commit from
upstream with additional changes and adjustments from
pragtical/pragtical#8

---------


* replace memory functions with SDL equivalent

---------

+++ src/api/renderer.c
@@ -242,7 +242,7 @@ static int f_font_get_size(lua_State *L) {
 static int f_font_set_size(lua_State *L) {
   RenFont* fonts[FONT_FALLBACK_MAX]; font_retrieve(L, fonts, 1);
   float size = luaL_checknumber(L, 2);
-  double scale = 1.0;
+  float scale = 1.0;
 #ifdef PRAGTICAL_USE_SDL_RENDERER
   RenWindow *window = ren_get_target_window();
   if (window != NULL) {
